Made OutlinedRectangle a local function instead of a member of the skin
Nothing outside the skin calls this, so it can easily be made local. But, the
reason it's worth making local is in case the default skin's drawing functions
are called by another skin: the other skin may not have an OutlinedRectangle,
and we still want to use the default one.